requirements.txt
aiogram==3.13.1
aiohttp==3.10.3
pydantic==2.11.0
pydantic-settings==2.4.0
python-dotenv==1.0.1

SQLAlchemy==2.0.32
aiosqlite==0.20.0
asyncpg==0.29.0
greenlet==3.0.3